Added wallet for users and admin views for appointments, transactions, wallet

+++ db/migrate/20220124135351_add_column_to_wallet.rb
@@ -1,5 +1,5 @@
 class AddColumnToWallet < ActiveRecord::Migration[6.1]
   def change
-    add_column :wallets, :current_balance, :decimal
+    add_column :wallets, :current_balance, :float
   end
 end
